backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
qrcode[pil]==7.4.2
apscheduler==3.10.4
twilio==8.11.1
pywebpush==1.14.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
opencv-python==4.9.0.80
websockets==12.0
